requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
alembic==1.13.1
redis==5.0.1
langchain==0.1.0
langchain-community==0.0.10
langchain-openai==0.0.2
pinecone-client==2.2.4
# Alternative: qdrant-client==1.6.9
sentence-transformers==2.2.2
pypdf2==3.0.1
python-docx==1.1.0
tiktoken==0.5.2
aiofiles==23.2.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
smtplib-ssl==1.0.4
pytest==7.4.3
pytest-asyncio==0.21.1
python-dotenv==1.0.0